Fixes error that would occur when searching for subtitles by hash if the size of the file was greater than 2^32 - 1 bytes.
First, xmlrpc was configured to properly handle long types (see
http://stackoverflow.com/questions/18533309/xmlrpc-response-datatype-long-possible).  Also, in
utils.File, the size attribute is now stored as a long.  Python actually unified int and long
awhile back (see PEP 237), so os.path.getsize actually returns an int regardless if it is larger
than a 32-bit number.  This commit casts the value returned by os.path.getsize to a long.  This
is nothing more than a quick fix to this particular bug.  It is possible that there could be more
bugs relating to integer size when using the xmlrpclib package.