fix #1 pip freeze svn requirements error
On successful install by pip install git+<url>, the freezed list
specified an error titles FIXME: Could not find a version that satisfies
the requirement. This was due to the svg repo configuration in setup.cfg
Removing that same removes the error. To create a successful install
and enable it to be ported via requirements.txt, use the -e option of
pip.